# Conflicts:
#	src/main/resources/static/css/grid.css
#	src/main/resources/static/script/board-community.js
#	src/main/webapp/WEB-INF/views/header.jsp

+++ src/main/resources/static/script/board-community.js
@@ -2,15 +2,28 @@ $(document).ready(function(){
 
 });
 
+function moveToUpdate(){
+    let boardNo = $('#board_num_hidden').val();
+
+    window.location.href='/board/community/'+boardNo+'/update';
+}
+
 function deleteBoard(){
-    $.ajax({
-        method:'DELETE',
-        url:`/board/community/delete/${boardNo}`,
-        contentType: 'application/json',
-        async:false
-    }).done(function (response){
-        console.log(response);
-    }).fail(function(){
-        console.log("게시글 삭제 실패");
-    })
+    // Trouble shooting : html 측에서 받아올 boardNo가 존재하지 않았음
+    let boardNo=$('#board_num_hidden').val();
+
+    if(confirm("게시글을 삭제하시겠습니까?")){
+        $.ajax({
+            method:'DELETE',
+            url:`/board/community/delete/${boardNo}`,
+            contentType: 'application/json',
+            async:false
+        }).done(function (response){
+            console.log(response);
+        }).fail(function(){
+            console.log("게시글 삭제 실패");
+        })
+    }else{
+        return false;
+    }
 }
